Wrapped storage functions into an IIFE
I reorganized the storage related code into an IIFE, so I could
encapsulate the theme key constant that was sitting at the top
of the file. I decided it made more sense to place it with the
functions that rely upon it instead of having it pollute the
global scope.