Setup the server's user route
I setup the route that will be responsible for providing the user data
to the front-end. I installed Octokit, GitHub's official Node.js SDK, and
used it to fetch information about users. After the information has been
received, either a success of an error message will be returned to the
client.